+++ src/commands.ts
@@ -1,21 +1,22 @@
 import * as vscode from 'vscode';
 
-import { GOTO_PACKAGE_COMMAND, GOTO_NPM_PAGE_COMMAND, PACKAGE_NOT_FOUND, NPM_URL } from './constants';
-import { getPackagePathByLine, getPackageNameByLine } from './fileManager';
+import { ExtensionCommand, ErrorMessage, NPM_URL, PACKAGE_JSON, QUICK_PICK_PLACEHOLDER, NODE_MODULES } from './constants';
+import { getPackageLines, getPackageNameByLine, getPackagePathByLine, openAndLocateFile, openBrowserPage, getPackagesNameDeeply, getPackageJSON } from './utils';
 
 export const gotoPackageCommand = vscode.commands.registerTextEditorCommand(
-    GOTO_PACKAGE_COMMAND,
+    ExtensionCommand.GOTO_PACKAGE,
     async (textEditor: vscode.TextEditor, edit: vscode.TextEditorEdit, lineNumber: number) => {
         try {
-            const { packagePath } = await checkPackageSelected(textEditor, lineNumber);
+            const document = textEditor.document;
+            const line = lineNumber || textEditor.selection.anchor.line || 0;
+            const packagePath = getPackagePathByLine(document, line);
 
-            await vscode.window.showTextDocument(
-                vscode.Uri.file(packagePath),
-                { preview: false }
-            );
-            vscode.commands.executeCommand(
-                'workbench.files.action.showActiveFileInExplorer'
-            );
+            if(packagePath) {
+                openAndLocateFile(packagePath);
+            }
+            else {
+                throw new Error(ErrorMessage.PACKAGE_NOT_FOUND);
+            }
         }
         catch(err) {
             handleError(err);
@@ -24,15 +25,19 @@ export const gotoPackageCommand = vscode.commands.registerTextEditorCommand(
 );
 
 export const gotoNpmPageCommand = vscode.commands.registerTextEditorCommand(
-    GOTO_NPM_PAGE_COMMAND,
+    ExtensionCommand.GOTO_NPM_PAGE,
     async (textEditor: vscode.TextEditor, edit: vscode.TextEditorEdit, lineNumber: number) => {
         try {
-            const { packageName } = await checkPackageSelected(textEditor, lineNumber);
+            const document = textEditor.document;
+            const line = lineNumber || textEditor.selection.anchor.line || 0;
+            const packageName = getPackageNameByLine(document, line);
 
-            vscode.commands.executeCommand(
-                'vscode.open',
-                vscode.Uri.parse(`${NPM_URL}${packageName}`)
-            );
+            if(packageName) {
+                openBrowserPage(`${NPM_URL}${packageName}`);
+            }
+            else {
+                throw new Error(ErrorMessage.PACKAGE_NOT_FOUND);
+            }
         }
         catch(err) {
             handleError(err);
@@ -40,26 +45,81 @@ export const gotoNpmPageCommand = vscode.commands.registerTextEditorCommand(
     }
 );
 
+export const gotoPickPackageCommand = vscode.commands.registerCommand(
+    ExtensionCommand.GOTO_PICK_PACKAGE,
+    async () => {
+        try {
+            const { rootPath } = vscode.workspace;
+            const searchNodeModules = vscode.workspace.getConfiguration().get('searchNodeModules');
 
-function checkPackageSelected(
-    textEditor: vscode.TextEditor,
-    lineNumber: number,
-) {
-    const document = textEditor.document;
-    const line = lineNumber || textEditor.selection.anchor.line || 0;
+            if(rootPath) {
+                const packagesPath: any = {};
+                let packagesName: null|string[] = null;
 
-    let packageName = getPackageNameByLine(document, line);
-    let packagePath = getPackagePathByLine(document, line);
+                if(searchNodeModules) {
+                    const nodeModulesPath = `${rootPath}/${NODE_MODULES}`;
 
-    if (packagePath) {
-        return Promise.resolve({
-            packageName,
-            packagePath
-        });
+                    packagesName = getPackagesNameDeeply(nodeModulesPath);
+                    packagesName.forEach(packageName => {
+                        packagesPath[packageName] = `${nodeModulesPath}/${getPackageJSON(packageName)}`;
+                    });
+                }
+                else {
+                    const packagePath = `${rootPath}/${PACKAGE_JSON}`;
+                    const document = await vscode.workspace.openTextDocument(
+                        vscode.Uri.file(packagePath)
+                    );
+
+                    packagesName = getPackageLines(document)
+                    .reduce((packagesName, line) => {
+                        const packageName = getPackageNameByLine(document, line.lineNumber);
+
+                        if(packageName) {
+                            const packagePath = getPackagePathByLine(document, line.lineNumber);
+                            packagesName.push(packageName);
+                            packagesPath[packageName] = packagePath;
+                        }
+
+                        return packagesName;
+                    }, [] as string[]);
+                }
+
+                const quickPickItems = (packagesName as string[]).map(packageName => {
+                    return {
+                        label: packageName,
+                        detail: packagesPath[packageName] ? '' : ErrorMessage.PACKAGE_NOT_FOUND,
+                    } as vscode.QuickPickItem;
+                });
+
+                const quickPick = vscode.window.createQuickPick();
+                quickPick.items = quickPickItems;
+                quickPick.placeholder = QUICK_PICK_PLACEHOLDER;
+
+                quickPick.onDidAccept(async () => {
+                    const packagePicked = quickPick.activeItems[0];
+                    const isPackageExist = !packagePicked.detail;
+
+                    if(isPackageExist) {
+                        const packageName = packagePicked.label;
+
+                        openAndLocateFile(packagesPath[packageName]);
+                        quickPick.hide();
+                        quickPick.dispose();
+                    }
+                    else {
+                        handleError(new Error(ErrorMessage.PACKAGE_NOT_FOUND));
+                    }
+                });
+                quickPick.show();
+            }
+            else {
+                throw new Error(ErrorMessage.WORKSPACE_IS_NOT_OPEN);
+            }
+        } catch(err) {
+            handleError(err);
+        }
     }
-
-    return Promise.reject(new Error(PACKAGE_NOT_FOUND));
-}
+);
 
 function handleError(err: Error) {
     vscode.window.showErrorMessage(err.message);