+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 All notable changes to the "goto-package" extension will be documented in this file.
 
+## [0.1.0] - 2019-03-09
+### Added
+- Add "Pick and go to a package" command to pick and go to selected package.
+
+### Changed
+- Rename commands and add gotoPackage prefix.
+
 ## [0.0.2] - 2019-01-30
 ### Added
 - Add "go to npm page" command to jump to npm page.